Added new sampling method
We are systematically sampling 350 samples from the 3540 total test examples in the Teco dataset for achieving 95% confidence level and 5% error margin. There are a total of 51 unique projects in the test set. So, the first strategy is to pick 7 samples ~(350/51) per project. For projects which have <=7 examples, we select all the examples from that project. For projects which have >7 examples, we randomly sample 7 samples from all examples of each such project. Through out the sampling process, we eliminate any examples that include assertThat or helper method based assertions. After executing the first strategy if we cannot reach the count of 350 samples, we randomly sample from the entire 3540 examples again to reach 350 sample count. All 350 samples are unique and do not include any example that contains  or  based assertions.